rag-backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
openai>=1.35.0
qdrant-client>=1.9.0
python-dotenv>=1.0.0
pydantic>=2.7.0
